+++ src/ts/components/fullscreentogglebutton.ts
@@ -28,17 +28,36 @@ export class FullscreenToggleButton extends ToggleButton<ToggleButtonConfig> {
       }
     };
 
+    let fullscreenAvailableHandler = () => {
+      if (player.isViewModeAvailable(player.exports.ViewMode.Fullscreen)) {
+        this.show();
+      } else {
+        this.hide();
+      }
+    };
+
+    uimanager.getConfig().events.onUpdated.subscribe(fullscreenAvailableHandler);
+
     player.on(player.exports.PlayerEvent.ViewModeChanged, fullscreenStateHandler);
 
     this.onClick.subscribe(() => {
-      if (player.getViewMode() === player.exports.ViewMode.Fullscreen) {
-        player.setViewMode(player.exports.ViewMode.Inline);
-      } else {
-        player.setViewMode(player.exports.ViewMode.Fullscreen);
+      if (!player.isViewModeAvailable(player.exports.ViewMode.Fullscreen)) {
+        if (console) {
+          console.log('Fullscreen unavailable');
+        }
+        return;
       }
+
+      const targetViewMode =
+        player.getViewMode() === player.exports.ViewMode.Fullscreen
+          ? player.exports.ViewMode.Inline
+          : player.exports.ViewMode.Fullscreen;
+
+      player.setViewMode(targetViewMode);
     });
 
     // Startup init
+    fullscreenAvailableHandler();
     fullscreenStateHandler();
   }
 }